+++ updown/acl_security.pl
@@ -20,22 +20,6 @@ print &ui_table_row($text{'acl_download'},
 		    [ 2, $text{'acl_nosched'} ],
 		    [ 0, $text{'no'} ] ]), 3);
 
-my $download_address_mode = $o->{'download_address_mode'};
-print &ui_table_row(
-	$text{'acl_download_addresses'} .
-		&ui_help($text{'acl_download_addresses_desc'}),
-	&ui_radio_table(
-		"download_address_mode", $download_address_mode,
-		[ [ 'all', $text{'acl_download_all'} ],
-		  [ 'public', $text{'acl_download_public'} ],
-		  [ 'listed', $text{'acl_download_listed'}."<br>".
-		    &ui_textarea(
-			"download_allowed_addresses",
-			join("\n", split(/\s+/,
-				$o->{'download_allowed_addresses'})),
-			4, 50) ] ]),
-	3);
-
 print &ui_table_row($text{'acl_users'},
 	&ui_radio_table("mode", $o->{'mode'},
 		[ [ 0, $text{'acl_all'} ],
@@ -61,12 +45,6 @@ my ($o) = @_;
 $o->{'upload'} = $in{'upload'};
 $o->{'max'} = $in{'max_def'} ? undef : $in{'max'};
 $o->{'download'} = $in{'download'};
-my $download_address_mode = $in{'download_address_mode'};
-$download_address_mode = 'public'
-	if ($download_address_mode !~ /^(public|listed|all)$/);
-$o->{'download_address_mode'} = $download_address_mode;
-$o->{'download_allowed_addresses'} = join(" ",
-	split(/\s+/, $in{'download_allowed_addresses'}));
 $o->{'mode'} = $in{'mode'};
 $o->{'users'} = $in{'mode'} == 0 || $in{'mode'} == 3 ? "" :
 		   $in{'mode'} == 1 ? $in{'userscan'}